+++ melosys-eessi-app/src/main/java/no/nav/melosys/eessi/integration/PersonFasade.java
@@ -2,7 +2,7 @@ package no.nav.melosys.eessi.integration;
 
 import java.util.List;
 
-import no.nav.melosys.eessi.integration.pdl.dto.sed.DnummerRekvisisjonTilMellomlagring;
+import no.nav.melosys.eessi.integration.pdl.web.identrekvisisjon.dto.DnummerRekvisisjonTilMellomlagring;
 import no.nav.melosys.eessi.models.person.PersonModell;
 import no.nav.melosys.eessi.service.personsok.PersonSokResponse;
 import no.nav.melosys.eessi.service.personsok.PersonsokKriterier;
