+++ melosys-eessi-app/src/main/java/no/nav/melosys/eessi/integration/pdl/PDLService.java
@@ -2,7 +2,7 @@ package no.nav.melosys.eessi.integration.pdl;
 
 import no.nav.melosys.eessi.integration.PersonFasade;
 import no.nav.melosys.eessi.integration.pdl.dto.*;
-import no.nav.melosys.eessi.integration.pdl.dto.sed.DnummerRekvisisjonTilMellomlagring;
+import no.nav.melosys.eessi.integration.pdl.web.identrekvisisjon.dto.DnummerRekvisisjonTilMellomlagring;
 import no.nav.melosys.eessi.models.exception.NotFoundException;
 import no.nav.melosys.eessi.models.person.PersonModell;
 import no.nav.melosys.eessi.models.person.UtenlandskId;
