+++ melosys-eessi-app/src/main/java/no/nav/melosys/eessi/integration/pdl/web/identrekvisisjon/dto/DnummerRekvisisjonNorskVegadresse.java
@@ -1,10 +1,10 @@
-package no.nav.melosys.eessi.integration.pdl.dto.sed;
+package no.nav.melosys.eessi.integration.pdl.web.identrekvisisjon.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
-import no.nav.melosys.eessi.integration.pdl.dto.sed.adresse.Vegadresse;
+import no.nav.melosys.eessi.integration.pdl.web.identrekvisisjon.dto.adresse.Vegadresse;
 
 @Data
 @Builder
