+++ melosys-eessi-app/src/main/java/no/nav/melosys/eessi/integration/pdl/web/identrekvisisjon/dto/adresse/VegadresseForPost.java
@@ -1,4 +1,4 @@
-package no.nav.melosys.eessi.integration.pdl.dto.sed.adresse;
+package no.nav.melosys.eessi.integration.pdl.web.identrekvisisjon.dto.adresse;
 
 import lombok.Builder;
 import lombok.Value;
